refactor: split scheduling and stats into dedicated files
